+++ novawallet/Common/Extension/SubstrateSdk/CallMetadata+TypeCheck.swift
@@ -3,10 +3,14 @@ import SubstrateSdk
 
 extension CallMetadata {
     func isArgumentTypeOf(_ name: String, closure: (String) -> Bool) -> Bool {
+        mapArgumentTypeOf(name, closure: { closure($0) }, defaultValue: false)
+    }
+
+    func mapArgumentTypeOf<T>(_ name: String, closure: (String) throws -> T, defaultValue: T) rethrows -> T {
         guard let argument = arguments.first(where: { $0.name == name }) else {
-            return false
+            return defaultValue
         }
 
-        return closure(argument.type)
+        return try closure(argument.type)
     }
 }
