+++ novawallet/Common/Services/ExtrinsicService/Substrate/BlockchainWeightFactory.swift
@@ -14,40 +14,50 @@ enum BlockchainWeightFactory {
         let argName: String
     }
 
-    static func convertExecuteCallWeightTypeToJson(
-        module: String,
-        codingFactory: RuntimeCoderFactoryProtocol,
-        weight: UInt64
-    ) throws -> JSON {
-        try convertCallVersionedWeightToJson(
-            for: .init(
-                path: .init(moduleName: module, callName: Xcm.executeCallName),
-                argName: Xcm.ExecuteCall<BigUInt>.CodingKeys.maxWeight.rawValue
-            ),
-            codingFactory: codingFactory,
-            weight: weight
-        )
-    }
-
-    static func convertCallVersionedWeightToJson(
+    static func convertCallVersionedWeightInWeightLimitToJson(
         for params: CallWeightConversionParams,
         codingFactory: RuntimeCoderFactoryProtocol,
         weight: UInt64
-    ) throws -> JSON {
+    ) throws -> JSON? {
         let context = codingFactory.createRuntimeJsonContext().toRawContext()
 
-        return try convertCallVersionedWeight(
-            for: params,
+        guard let callType = codingFactory.getCall(for: params.path) else {
+            return nil
+        }
+
+        let optWeightType: String? = callType.mapArgumentTypeOf(
+            params.argName,
+            closure: { weightLimitType in
+                guard let weightLimitNode = codingFactory.getTypeNode(for: weightLimitType) as? SiVariantNode else {
+                    return nil
+                }
+
+                return weightLimitNode.typeMapping
+                    .first { $0.name == Xcm.WeightLimitFields.limited }?
+                    .node.typeName
+            },
+            defaultValue: nil
+        )
+
+        guard let weightType = optWeightType else {
+            return nil
+        }
+
+        return try convertTypeVersionedWeight(
+            for: weightType,
             codingFactory: codingFactory,
             handlers: .init(
                 v1Handler: { _, weight in
                     try weight.toScaleCompatibleJSON(with: context)
-                }, v1P5Handler: { _, weight in
+                },
+                v1P5Handler: { _, weight in
                     try weight.toScaleCompatibleJSON(with: context)
-                }, v2Handler: { _, weight in
+                },
+                v2Handler: { _, weight in
                     try weight.toScaleCompatibleJSON(with: context)
                 }
-            ), builder: .null,
+            ),
+            builder: JSON.null,
             weight: weight
         )
     }
@@ -63,18 +73,36 @@ enum BlockchainWeightFactory {
             return builder
         }
 
-        // v1 require only uint64 weight
-        let isV1 = callType.isArgumentTypeOf(params.argName) { argumentType in
-            codingFactory.isUInt64Type(argumentType)
-        }
+        return try callType.mapArgumentTypeOf(
+            params.argName,
+            closure: { argumentType in
+                try Self.convertTypeVersionedWeight(
+                    for: argumentType,
+                    codingFactory: codingFactory,
+                    handlers: handlers,
+                    builder: builder,
+                    weight: weight
+                )
+            },
+            defaultValue: builder
+        )
+    }
+
+    static func convertTypeVersionedWeight<T>(
+        for type: String,
+        codingFactory: RuntimeCoderFactoryProtocol,
+        handlers: CallWeightConversionHandlers<T>,
+        builder: T,
+        weight: UInt64
+    ) throws -> T {
+        // v1 require only uint64 or compact number weight
+        let isV1 = codingFactory.isUInt64Type(type) || codingFactory.isCompactType(type)
 
         if isV1 {
             return try handlers.v1Handler(builder, .init(value: weight))
         } else {
-            // verision 1.5 contains only 1 field and v2 contains 2 fields
-            let isV1P5 = callType.isArgumentTypeOf(params.argName) { argumentType in
-                codingFactory.isStructHasFieldsCount(argumentType, count: 1)
-            }
+            // version 1.5 contains only 1 field and v2 contains 2 fields
+            let isV1P5 = codingFactory.isStructHasFieldsCount(type, count: 1)
 
             if isV1P5 {
                 return try handlers.v1P5Handler(builder, .init(refTime: weight))
