+++ novawallet/Common/Substrate/Types/Xcm/XcmPalletMetadataQueryFactory.swift
@@ -3,6 +3,10 @@ import RobinHood
 import SubstrateSdk
 
 protocol XcmPalletMetadataQueryFactoryProtocol {
+    func createModuleNameResolutionWrapper(
+        for runtimeProvider: RuntimeProviderProtocol
+    ) -> CompoundOperationWrapper<String>
+
     func createLowestMultiassetsVersionWrapper(
         for runtimeProvider: RuntimeProviderProtocol
     ) -> CompoundOperationWrapper<Xcm.Version?>
@@ -57,6 +61,30 @@ extension XcmPalletMetadataQueryFactory: XcmPalletMetadataQueryFactoryProtocol {
         createXcmTypeVersionWrapper(
             for: runtimeProvider,
             typeName: "xcm.VersionedMultiLocation"
+        )
+    }
+
+    func createModuleNameResolutionWrapper(
+        for runtimeProvider: RuntimeProviderProtocol
+    ) -> CompoundOperationWrapper<String> {
+        let coderFactoryOperation = runtimeProvider.fetchCoderFactoryOperation()
+
+        let moduleResolutionOperation = ClosureOperation<String> {
+            let metadata = try coderFactoryOperation.extractNoCancellableResultData().metadata
+            guard let moduleName = Xcm.possibleModuleNames.first(
+                where: { metadata.getModuleIndex($0) != nil }
+            ) else {
+                throw XcmTransferServiceError.noXcmPalletFound
+            }
+
+            return moduleName
+        }
+
+        moduleResolutionOperation.addDependency(coderFactoryOperation)
+
+        return CompoundOperationWrapper(
+            targetOperation: moduleResolutionOperation,
+            dependencies: [coderFactoryOperation]
         )
     }
 }
